EDHM v21.00 for FDev U4.2.0.2
- First Fix for Elite Dangerous v4.2.0.2 (Vanguards Update)
  - Fix Orange Elements and other problems caused by Vanguards Update
  - NEW: New Squadron Menu Color Mode (Cyan Elite Default or XML) (y229)
  - NEW: The HoloMe Menu is now modded and therefore takes on the XML Color

If the New Squadron Menu mod turns out to be problematic, it may be changed or even removed in a future update.

NOTE: Since I don't have access to a Squadron Carrier (SFC), it's impossible for me to mod SFC-related items, so you might find some unmodded items in your SFC menus. These menus may be modded in the future if I obtain an SFC.